refactor(mpp): extract partitioner module + delete legacy MppShuffleExec
Splits the still-useful hash-routing primitives out of the soon-to-be-
deleted `mpp::shuffle` module into a dedicated `mpp::partitioner` module,
then removes `mpp::shuffle` entirely.

What stays (now in `mpp/partitioner.rs`):
- `RowPartitioner` trait
- `HashPartitioner` (production hash routing, ahash 0/0/0/0 seed for
  REPARTITION_RANDOM_STATE compatibility)
- `FixedTargetPartitioner` (scalar-final gather)
- `split_batch_by_partition` helper
- The corresponding unit tests + the `ModuloPartitioner` test helper.
- `cfg(test)` `pub mod tests {}` so future plan-build tests can re-use
  `ModuloPartitioner` and `sample_batch`.

What goes:
- `MppShuffleExec` and its peer-to-peer producer/consumer wiring
  (`ShuffleWiring`, `ShuffleState`, `build_shuffle_stream`,
  `log_shuffle_eof`, `run_shuffle_pump`).
- The 6 `pump_*` integration tests covering that wiring (-7 tests; the
  remaining 11 partitioner tests are preserved verbatim).

Net diff: -1167 LOC (-1500 deleted in `shuffle.rs`, +382 added in
`partitioner.rs`). 51 `mpp::` lib tests still pass (51 = 57 - 6 legacy
pump tests); clippy + cargo fmt clean.

The producer-side `ExecutionPlan` for the new coordinator/worker
architecture is `mpp::shm_mq_producer::ShmMqProducerExec` (introduced in
the previous PR). The walker that emits it + the customscan role split
+ end-to-end bench validation land in a follow-up PR.
