+++ pg_search/src/postgres/customscan/mpp/mod.rs
@@ -26,11 +26,11 @@
 //! consumer-side backpressure from producer-side backpressure.
 
 pub mod mesh;
+pub mod partitioner;
 pub mod rpc_mesh;
 pub mod session;
 pub mod shm_mq_producer;
 pub mod shm_mq_transport;
-pub mod shuffle;
 pub mod transport;
 pub mod worker;
